Authorize pundit

+++ db/migrate/20240920122510_add_best_answer_to_questions.rb
@@ -1,6 +1,6 @@
 class AddBestAnswerToQuestions < ActiveRecord::Migration[7.2]
   def change
-    add_column :questions, :best_answer_id, :integer
+    add_column :questions, :best_answer_id, :bigint
     add_foreign_key :questions, :answers, column: :best_answer_id
 
   end
